+++ ZipUtilities/NSData+NOZAdditions.m
@@ -35,8 +35,14 @@
 
 - (NSData *)noz_dataByCompressing:(id<NOZEncoder>)encoder compressionLevel:(NOZCompressionLevel)compressionLevel
 {
+    if (self.length == 0) {
+        return nil;
+    }
+
     __block NSMutableData *encodedData = [NSMutableData data];
-    id<NOZEncoderContext> context =
+
+    @autoreleasepool {
+        id<NOZEncoderContext> context =
         [encoder createContextWithBitFlags:0
                           compressionLevel:compressionLevel
                              flushCallback:^BOOL(id<NOZEncoder> callbackEncoder, id<NOZEncoderContext> encoderContext, const Byte *bufferToFlush, size_t length) {
@@ -44,20 +50,21 @@
                                  return YES;
                              }];
 
-    if (![encoder initializeEncoderContext:context]) {
-        return nil;
-    }
-
-    __block BOOL wasError = NO;
-    [self enumerateByteRangesUsingBlock:^(const void *bytes, NSRange byteRange, BOOL *stop) {
-        wasError = ![encoder encodeBytes:(const Byte *)bytes length:byteRange.length context:context];
-        if (wasError) {
-            *stop = YES;
+        if (![encoder initializeEncoderContext:context]) {
+            return nil;
         }
-    }];
 
-    if (wasError || ![encoder finalizeEncoderContext:context]) {
-        return nil;
+        __block BOOL wasError = NO;
+        [self enumerateByteRangesUsingBlock:^(const void *bytes, NSRange byteRange, BOOL *stop) {
+            wasError = ![encoder encodeBytes:(const Byte *)bytes length:byteRange.length context:context];
+            if (wasError) {
+                *stop = YES;
+            }
+        }];
+
+        if (wasError || ![encoder finalizeEncoderContext:context]) {
+            return nil;
+        }
     }
 
     return encodedData;
@@ -65,39 +72,46 @@
 
 - (NSData *)noz_dataByDecompressing:(id<NOZDecoder>)decoder
 {
-    __block NSMutableData *encodedData = [NSMutableData data];
-    id<NOZDecoderContext> context =
+    if (self.length == 0) {
+        return nil;
+    }
+
+     __block NSMutableData *decodedData = [NSMutableData data];
+
+    @autoreleasepool {
+        id<NOZDecoderContext> context =
         [decoder createContextForDecodingWithBitFlags:0
                                         flushCallback:^BOOL(id<NOZDecoder> callbackDecoder, id<NOZDecoderContext> decoderContext, const Byte *bufferToFlush, size_t length) {
-                                            [encodedData appendBytes:bufferToFlush length:length];
+                                            [decodedData appendBytes:bufferToFlush length:length];
                                             return YES;
                                         }];
 
-    if (![decoder initializeDecoderContext:context]) {
-        return nil;
-    }
-
-    __block BOOL wasError = NO;
-    [self enumerateByteRangesUsingBlock:^(const void *bytes, NSRange byteRange, BOOL *stop) {
-        wasError = ![decoder decodeBytes:(const Byte *)bytes length:byteRange.length context:context];
-        if (wasError) {
-            *stop = YES;
+        if (![decoder initializeDecoderContext:context]) {
+            return nil;
         }
-    }];
 
-    while (!wasError && !context.hasFinished) {
-        wasError = ![decoder decodeBytes:NULL length:0 context:context];
+        __block BOOL wasError = NO;
+        [self enumerateByteRangesUsingBlock:^(const void *bytes, NSRange byteRange, BOOL *stop) {
+            wasError = ![decoder decodeBytes:(const Byte *)bytes length:byteRange.length context:context];
+            if (wasError) {
+                *stop = YES;
+            }
+        }];
+
+        while (!wasError && !context.hasFinished) {
+            wasError = ![decoder decodeBytes:NULL length:0 context:context];
+        }
+
+        if (wasError) {
+            return nil;
+        }
+        
+        if (![decoder finalizeDecoderContext:context]) {
+            return nil;
+        }
     }
 
-    if (wasError) {
-        return nil;
-    }
-
-    if (![decoder finalizeDecoderContext:context]) {
-        return nil;
-    }
-
-    return encodedData;
+    return decodedData;
 }
 
 @end